refactor: rearrange Community related files
