+++ src/components/community/CommunityCard.js
