+++ src/pages/Post.js
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from "react";
 import Styled from "styled-components";
-import PostCard from "../components/PostCard";
+import PostCard from "../components/community/PostCard";
 import { Link } from "react-router-dom";
 import Navbar from "../components/Navbar";
 import { getPostList } from "../lib/api/post";
